Restored the app and refactored it into a package.
